Simple Magic round 3 slice 1: paying more can no longer buy better odds
was: SkillChecks.successPermille built its raw permille in int --
base + 10 * (score - resist). Every resist the shipped shelf produces is a
two-digit number, so nothing noticed; but the vocabulary deliberately puts NO
ceiling on magnitude for the temperature and vitality axes, on the argument
that an outsized crafting prices itself out instead of being refused. Past a
resist of ~215 million that multiply wrapped POSITIVE and the argument
inverted: the most expensive crafting authorable was the EASIEST one in the
ward to land, pinned at the ceiling, with the check line printing the honest
resist beside the dishonest percentage.

EffectComponent.transferPoints could wrap on its own too (a big magnitude
times a long duration), and a wrapped transfer count is a NEGATIVE resist --
the same lie arriving one step earlier.

now: the whole chain is long and saturating. transferPoints in long from the
first operation, SpellDefinition sums in long and saturates at
SpellCost.RESIST_CEILING, resistFor computes and clamps in long, and
successPermille has a long-resist overload so the one threshold function every
check family shares cannot wrap for any of them.

new: CostModelMonotonicityTest sweeps magnitude across the whole legal range
on BOTH signs (1..12, powers of two to 2^30, both sides of the old wrap point,
MAX_VALUE and MIN_VALUE), plus duration to the tick ceiling, distance to 4096
tiles and spread to radius 64 -- asserting at every step that difficulty never
falls and odds never rise. The biggest crafting a raws author can type now
sits on the check's FLOOR, which is where it always claimed to sit.

also (minor from review): the area axis paid no unbridged tax. A spread
reaches bodies the caster is not touching, which is exactly what L459 gates
behind the gift, so a TOUCH crafting spreading five tiles cost 30 while the
single-target bolt reaching that far cost 40 -- the crowd version was the
bargain. Both unbridged shapes pay the tax now, once each, and the test pins
that a spread never undercuts the bolt at the same radius. No shipped spell
authors an area, so no shipped difficulty moved.

+++ client-observer/src/main/java/com/trojia/client/inspect/CheckLineFormatter.java
@@ -181,7 +181,7 @@ public final class CheckLineFormatter {
      * @param resist    the resist the sim rolled against ({@code SpellCost.resistFor})
      */
     public static String spellLine(SkillTrackRegistry tracks, int casterId, int skillRaw,
-            String spellName, int resist, boolean worked) {
+            String spellName, long resist, boolean worked) {
         String outcome = worked ? "THE LINK HOLDS" : "THE LINK SLIPS";
         if (!tracks.isWired() || skillRaw == Actor.NONE) {
             return "[crafting -- " + outcome + "]";
